Adding support for input actions
This change adds a way to enable the use of input map. On top of the current way of retrieving data, this allows you to use Godot's input system. By filling the 4 labels for the input actions of up, right, bottom and left, this will emulate the use of a true joystick. This is extremely useful when you are making a game cross platform and don't want to change the code for each input system. You could have this virtual joystick trigger the same input action a real one would.